Updated Various Sounds
Adjusted timing, distance decay and volume of both environment and puzzle sounds to fit better into the mix. Updated the player and enemy attack sounds as well as the player block sound (Still need to fix up the damage received sounds) I also changed one of the prefabs to reference the correct sound event. There's still a lot of sounds which haven't been implemented into the game which I've already created, happy to help work through the implementations if needed